# openapi/flora.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "flora API",
    "description": "Manages @Flora guild deployments, User Bots, Server Custom Bots, authentication, builds, secrets, key-value stores, logs, and operator orchestration.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "flora API base path"
    }
  ],
  "paths": {
...
    "/builds/": {
      "post": {
        "tags": ["Builds"],
        "summary": "Create a build",
        "description": "Queues a server-side build for the provided project archive.",
        "operationId": "create_build_handler",
        "responses": {
          "200": {
            "description": "Build queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBuildResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/builds/{build_id}": {
      "get": {
        "tags": ["Builds"],
        "summary": "Get build status",
        "description": "Returns the current status, logs, and artifacts for a build.",
        "operationId": "get_build_handler",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "description": "Build ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Build status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Build not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/builds/{build_id}/logs": {
      "get": {
        "tags": ["Builds"],
        "summary": "Stream build logs",
        "description": "Streams build output as Server-Sent Events until the upstream build stream closes. The authenticated caller must have access to the build's guild or User Bot scope.",
        "operationId": "stream_build_logs_handler",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "description": "Build ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Event stream containing build output",
            "content": {
              "text/event-stream": {}
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuildArtifactResponse": {
        "type": "object",
        "description": "Build artifact output paths.",
        "required": ["bundle", "source_map"],
        "properties": {
          "bundle": {
            "type": "string",
            "description": "Bundled JavaScript output."
          },
          "source_map": {
            "type": "string",
            "description": "Source map for the bundle."
          }
        }
      },
      "BuildStatusResponse": {
        "type": "object",
        "description": "Current build status and output.",
        "required": ["build_id", "guild_id", "entry", "status", "logs"],
        "properties": {
          "artifact": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BuildArtifactResponse",
                "description": "Generated build artifacts, if available."
              }
            ]
          },
          "build_id": {
            "type": "string",
            "description": "Build identifier."
          },
          "entry": {
            "type": "string",
            "description": "Entry file path used for the build."
          },
          "error": {
            "type": ["string", "null"],
            "description": "Error detail when the build fails."
          },
          "finished_at": {
            "type": ["string", "null"],
            "description": "Build completion time in RFC3339 (UTC)."
          },
          "guild_id": {
            "type": "string",
            "description": "Guild ID the build targets."
          },
          "logs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Build logs, newest last."
          },
          "started_at": {
            "type": ["string", "null"],
            "description": "Build start time in RFC3339 (UTC)."
          },
          "status": {
            "type": "string",
            "description": "Current build status string."
          }
        }
      },
      "CreateBuildResponse": {
        "type": "object",
        "description": "Response returned when a build is created.",
        "required": ["build_id", "status"],
        "properties": {
          "build_id": {
            "type": "string",
            "description": "Build identifier."
          },
          "status": {
            "type": "string",
            "description": "Current build status string."
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status", "detail", "instance", "request_id"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Actionable explanation of this specific failure occurrence."
          },
          "instance": {
            "type": "string",
            "description": "Request path identifying where this failure occurred."
          },
          "request_id": {
            "type": "string",
            "description": "Correlation identifier also returned in the `X-Request-ID` header."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code returned for this failure occurrence.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Stable, human-readable summary of the failure category."
          },
          "type": {
            "type": "string",
            "description": "Stable relative URI identifying the category of failure."
          },
          "violations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemViolation"
            },
            "description": "Field-level validation failures, when applicable."
          }
        }
      },
      "ProblemViolation": {
        "type": "object",
        "required": ["field", "message"],
        "properties": {
          "field": {
            "type": "string",
            "description": "Identifies the invalid request field using its JSON or parameter path."
          },
          "message": {
            "type": "string",
            "description": "Explains how to correct the invalid field."
          }
        }
      },
...
    }
  },
...
}